The integer timeout is cast to a void *, passed to this function, and cast back.
Add a cast to (long) to quiet the compiler


git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@138094 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/sqlite/libsqlite/src/main.c
@@ -764,7 +764,8 @@ static int sqliteDefaultBusyCallback(
 #if SQLITE_MIN_SLEEP_MS==1
   int delay = 10;
   int prior_delay = 0;
-  int timeout = (int)Timeout;
+  /* We seem to be called by a generic cb mechanism which passes void ptrs */
+  int timeout = (int)(long)Timeout;
   int i;
 
   for(i=1; i<count; i++){ 
@@ -783,7 +784,7 @@ static int sqliteDefaultBusyCallback(
   sqliteOsSleep(delay);
   return 1;
 #else
-  int timeout = (int)Timeout;
+  int timeout = (int)(long)Timeout;
   if( (count+1)*1000 > timeout ){
     return 0;
   }
